agent-runtime/app/requirements.txt
boto3>=1.34.0
pydantic>=2.0.0,<3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aws-opentelemetry-distro>=0.10.0
opentelemetry-api>=1.20.0
